# Enable async for io_uring reader and disable sqpoll