--- a/accounts-db/src/io_uring/sequential_file_reader.rs
+++ b/accounts-db/src/io_uring/sequential_file_reader.rs
@@ -348,6 +348,7 @@ impl RingOp<SequentialFileReaderState> for ReadOp {
         .offset(*file_off as u64)
         .ioprio(IO_PRIO_BE_HIGHEST)
         .build()
+        .flags(squeue::Flags::ASYNC)
     }
 
     fn complete(