#### Problem
* sequential file reader is submitting IOs in so called non-blocking mode (as opposed to async), which makes them executed primarily on the single thread unless they are recognized as blocking
* the above is currently offset by use of sqpoll, which uses dedicated kernel thread spinning full to consume submitted ops and executing them

Primary observation (from a few tests) is that `ASYNC` is always beneficial, IOs are immediately spread across workers, which increases parallelism and any overhead of handing ops to and back from worker seems negligible.
In case of sequential file reader this also makes sqpoll optional - it still does slightly help with avoiding user thread calling `submit` (i.e. syscall), but with the right sizing of submission and completion queues, the impact is very small. The downside of sqpoll is that it utilizes the whole additional core, which make it bad for scaling up use of reader (e.g. reading multiple files)
 
#### Summary of Changes
* enable ASYNC
* adjust io_uring sizing to properly amortize submits - allow as many ops as buffers to be in-flight, make submissions queue sized in such a way, that after filling half of it we will perform `submit`

Note: one side-effect in case of snapshot unpack is that reader and file-creator (running in the same user-thread) now share the kernel worker pool, with a max size specified by both (4) being used - that way read ops end up being placed on more worker threads and all ops are being executed together.

##### Performance impact
There is a very tiny regression (disabled sqpoll) in snapshot unpack times (it's within variance, but avg is a bit higher):

master:
91.8s, 91.2s, 91.8s, 92.9s

PR:
91.4s, 92.0s, 94.1s, 92.5s